feat: simulateTxBundle function working
